feat: CPU software decode backends (edge264 H.264, hevc.js H.265)
Add two CPU decode backends behind the unified 'sw' backend, each a
vendored, statically-compiled C/C++ core with an MIT-compatible license:

- vacc-sw-decode: edge264 (C) for H.264
- vacc-software-decode: hevc.js (C++) for H.265. The Rust layer keeps
  bitstream parsing, DPB/POC tracking and display-order reordering;
  the C++ core only reconstructs pixels.

The 'decode' example gets per-backend cargo features (edge264, hevcjs,
nvdec, vaapi, vulkan; all on by default).

H.265 samples decode byte-exact vs FFmpeg (300 frames each, including
the multi-slice + WPP stress stream).

+++ crates/vacc-nvdec-decode/src/vp9.rs
@@ -186,7 +186,7 @@ pub fn build_cuvid_vp9_picparams(
         for seg in 0..8 {
             for f in 0..4 {
                 vp9.segmentFeatureEnable[seg][f] = ((sg.feature_enabled[seg] >> f) & 1) as c_uchar;
-                vp9.segmentFeatureData[seg][f] = sg.feature_data[seg][f] as i16;
+                vp9.segmentFeatureData[seg][f] = sg.feature_data[seg][f];
             }
         }
         vp9.mb_segment_tree_probs = sg.segmentation_tree_probs;
